+++ app/src/main/java/com/simenko/qmapp/ui/main/structure/forms/subdepartment/SubDepartmentForm.kt
@@ -91,8 +91,8 @@ fun SubDepartmentForm(
             InfoLine(modifier = modifier.padding(start = 0.dp), title = "Department", body = concatTwoStrings(sdComplete.department.depAbbr, sdComplete.department.depName))
             Spacer(modifier = Modifier.height(10.dp))
             RecordFieldItem(
-                valueParam = Triple(sdComplete.subDepartment.subDepOrder.let { if (it == NoRecord.num) EmptyString.str else it }.toString(), fillInErrors.subDepartmentOrderError) {
-                    viewModel.setSubDepartmentOrder(if(it == EmptyString.str) NoRecord.num else it.toInt())
+                valueParam = Triple(sdComplete.subDepartment.subDepOrder.let { if (it == NoRecord.num.toInt()) EmptyString.str else it }.toString(), fillInErrors.subDepartmentOrderError) {
+                    viewModel.setSubDepartmentOrder(if(it == EmptyString.str) NoRecord.num.toInt() else it.toInt())
                 },
                 keyboardNavigation = Pair(orderFR) { abbreviationFR.requestFocus() },
                 keyBoardTypeAction = Pair(KeyboardType.Number, ImeAction.Next),
